Added 'beta' and 'prod' flavours. Beta flavor always points to EU for now.
